mempool: Remove ancestor/descendant related stats from RPC
Summary:
```
The mempool quadratic stats themselves will soon be removed in a future commit.
However, we are removing them from RPC now in this commit to keep things
manageable.

This commit removes the following stats from RPC:

- "ancestor", "descendant", "descendantcount", "descendantsize", "descendantfees",
  "ancestorcount", "ancestorsize", "ancestorfees"
- RPC's affected: `getrawmempool true`, `getmempoolancestors`,
  `getmempooldescendants`, `getmempoolentry`
```

Port of [[https://gitlab.com/bitcoin-cash-node/bitcoin-cash-node/merge_requests/1129 | bchn#1129]].

Depends on D13133 and D13136.

Instead of removing the fields, we deprecate them for backward compatibility.
The mempool_packages.py and mempool_updatefromblock.py tests are kept
unchanged with the deprecation option so we are still testing the feature.
It's also easier than adding the cases to the deprecated RPCs functional test.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Subscribers: PiRK

Differential Revision: https://reviews.bitcoinabc.org/D13137