p2p: update inbound eviction protection for multiple networks, add I2P peers
Summary:
This pull updates and abstracts our inbound eviction protection to make it fully ratio-based and easily extensible to peers connected via high-latency privacy networks that we newly support, like I2P and perhaps others soon, as these peers are disadvantaged by the latency criteria of our eviction logic.

It then adds eviction protection for peers connected over I2P. We've observed over the past few months that I2P peers have a min ping latency similar to or greater than that of onion peers.

The algorithm is a basically a multi-pass knapsack:
 - Count the number of eviction candidates in each of the disadvantaged privacy networks.
 - Sort the networks from lower to higher candidate counts, so that a network with fewer candidates will have the first opportunity for any unused slots remaining from the previous iteration. In the case of a tie in candidate counts, priority is given by array member order from first to last, guesstimated to favor more unusual networks.
 - Iterate through the networks in this order. On each iteration, allocate each network an equal number of protected slots targeting a total number of candidates to protect, provided any slots remain in the knapsack.
 - Protect the candidates in that network having the longest uptime, if any in that network are present.
 - Continue iterating as long as we have non-allocated slots remaining and candidates available to protect.

The goal of this logic is to promote the diversity of our peer connections.

This is a backport of [[https://github.com/bitcoin/bitcoin/pull/21261 | core#21261]]

Depends on D11042

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D11051