mempool: Slight performance nit when erasing from setEntries
Summary:
```
When re-reading all of the mempool code in master I noticed that in 2 places, a working `setEntries` (which is an alias for a
`std::set<CTxMemPool::txiter>`) was being erased-from inefficiently.

In these two places we are grabbing the first `txiter` from the set then we are erasing it.  In the extant code, the `txiter` is being passed as a *key* to the set. Instead, we should be erasing using the slightly more efficient `std::set(std::set::iterator)` overload, which does not need to do any O(log N) lookups.
```

Port of [[https://gitlab.com/bitcoin-cash-node/bitcoin-cash-node/merge_requests/1175 | bchn#1175]].
Depends on D13171.

There is no change in behavior.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D13176