refactor: Remove chainparams arg from CChainState member functions
Summary:
Passing this is confusing and redundant with the m_params member.
This is a backport of [[https://github.com/bitcoin/bitcoin/pull/21789 | core#21789]] [2/2]
https://github.com/bitcoin/bitcoin/pull/21789/commits/fa0d9211ef87a682573aaae932c0c440acbcb8a8

Depends on D12189

The result is not quite as spectacular as in the source material, as for many of these methods the Bitcoin ABC codebases does not pass directly a CChainParams object but rather a Config object. I will attempt to do the same operation for Config in a separate diff.

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D12190