[Cashtab] Reduce bundle size
Summary:
Revert to `unpkg` hosted `minimal-slp-wallet` for the web app, as the dependency is large and may be causing long load times at cashtabapp.com. Also upgrade antd/icons to latest version, which is smaller.

The extension will still need to use the module, but load times are not an issue there as the user installs them. The next diff bumping the extension version will handle this distinction between the webapp and the extension. As the extension can be expected to require unique dependencies in the future, might as well handle it on this one.

Test Plan: `npm start`, confirm can send a BCHA tx, confirm can send an SLPA tx

Reviewers: #bitcoin_abc, majcosta

Reviewed By: #bitcoin_abc, majcosta

Differential Revision: https://reviews.bitcoinabc.org/D9460