[Cashtab] Patch z-index css issue in forms vs footer
Summary: User reported issue where forms appear on top of footer if user has form selected. This increases the z-index parameter of the `Footer` component to ensure it is always on top of forms.

Test Plan: `npm start`, navigate to Tokens in mobile view, adjust height so that Create Token form covers full height of page, select lowest form, scroll down and observe form is not on top of nav panel

Reviewers: #bitcoin_abc, majcosta

Reviewed By: #bitcoin_abc, majcosta

Differential Revision: https://reviews.bitcoinabc.org/D9484