[avalanche] don't ignore avapoll and avaresponse while importing
Summary:
It is better not to ignore avaresponses while importing because it prevents detection and discouragement of invalid responses.
For avapolls, it is better to respond that we don't know than leave the peer hanging.

Add a test for the discouragement behavior of  avaresponses that do not correspond to previous avapolls.

Addresses review D9482#inline-46570

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D9494