[Cashtab] Change dust param from BCHA to sats
Summary:
The `currency.dust` param has been set to `'0.00000546'` to reflect the minimum amount of BCHA that the app will send.

To support potential decimal place moves, this figure should be in `sats`.

This diff deletes the `currency.dust` parameter, replaces it with `currency.dustSats`, and parses this value for the arbitrary `currency.cashDecimals` figure wherever it is used in the app. Unit tests are included to ensure consistent application.

Test Plan:
`npm test`
`npm start`
Try to send 0.00000545 BCHA and note error msg is correct

Reviewers: #bitcoin_abc, majcosta

Reviewed By: #bitcoin_abc, majcosta

Differential Revision: https://reviews.bitcoinabc.org/D9495