refactor: Mark CAddrMan::Select and GetAddr const
Summary:
```
To clarify that a call to this only changes the random state and nothing else.
```

Backport of [[https://github.com/bitcoin/bitcoin/pull/21940 | core#21940]].

The source material does more that what the title says, but only the 2 first commits are relevant to our codebase, making it actually accurate:
 - https://github.com/bitcoin/bitcoin/pull/21940/commits/fa02934c8c9d290ea4d12683e8680c70967a4d3a
 - https://github.com/bitcoin/bitcoin/pull/21940/commits/fae0c79351ce34186249d44af0c5c9c7521f4b6c

This is due to the fact that we didn't backport the fuzzer yet, and the addrman deterministic behavior has been changed in D11827 and the locks properly held in the previous backports.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D12265