[avalanche] change how the proof id is computed
Summary:
Do a first hash of the proof's data without the master key, then append the master key to this hash and hash again. Fix all tests that depend on the proofid.

This is a first step in creating a LimitedProofId without the master key, to be used in the avahello message to verify a delegation without knowing a proof.
For ease of review, I split the change in two diffs.

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Subscribers: deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D9543