Make the implementation details of TxRequestTracker use an uint256
Summary:
Instead of a TxId. The public interfaces are left unchanged. This is a
step toward making it a template class able to handle both transactions
and proofs, which are uint256 subclasses.

There is a tweak to make the type conversion from the `GetRequestable`
public interface, which involves updating the signature of the internal
implementation.

Test Plan:
  ninja all check-all
  ./test/functional/test_runner.py p2p_tx_download
  ninja bitcoin-fuzzers

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Maniphest Tasks: T1611

Differential Revision: https://reviews.bitcoinabc.org/D9566