Add missing <optional> headers
Summary:
This is a leftover from D8494: when switching from boost::optional to
std::optional the headers were not updated.

Also added a missing <limits> include in the FuzzDataProvider.h that
supposedly was pulled indirectly. Note that <climits> cannot be removed
due to CHAR_BIT.

Test Plan:
On an up-to-date archlinux:
  ninja all check
  ninja bitcoin-fuzzers

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D9576