[avalanche] Add an API to store and retrieve a proof
Summary:
We will need to store proofs that we downloaded and retrieve them for
e.g. relaying. The proof storage is done by the peer manager, which acts
as the valid proof pool and creates a new peer for each new proof. The
interface is provided by the processor so the peer manager remains
private.

Depends on D9617.

Test Plan:
  ninja check-avalanche

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Subscribers: deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D9595