Validate port-options
Summary:
Check `port` options for invalid values (ports are parsed as uint16, so
in practice values >65535 are invalid; port 0 is undefined and therefore
considered invalid too). This allows for an early rejection of faulty
values and an supplying an informative message to the user.

Splits tests in `feature_proxy.py` to cover both invalid `hostname`
and `port` values.

Adds a release-note as previously valid `-port` and `-rpcport` values
can now result in errors.

This is a backport of [[https://github.com/bitcoin/bitcoin/pull/22087 | core#22087]] and a partial backport of [[https://github.com/bitcoin/bitcoin/pull/24205 | core#24205]]
https://github.com/bitcoin/bitcoin/pull/24205/changes/d5edb087082a50e6f7d413c3b43fdf1e6a20d29b
https://github.com/bitcoin/bitcoin/pull/24205/changes/8332e6e4cf45455fea0bf1f7527256cdb7bb1e6d

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D19730