[avalanche] rename parameters in DelegationBuilder::addLevel
Summary:
There are multiple keys called master in the avalanche codebase: the
proof's master key, the last level of the delegation and the
`-avamasterkey` command line argument.

This reduces the risk of confusion by renaming `master` into
`delegatedPubKey` in the delegation builder, prior to adding a new
member `proofMaster` to that class.

This does not change any behavior.

Test Plan: `ninja`

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D9597