test: run mempool_reorg.py even with wallet disabled
Summary:
```
- run mempool_reorg.py even when the wallet is not compiled
- add `locktime` argument to `create_self_transfer` and `send_self_transfer`
- use more logs instead of comments
```

Backport of [[https://github.com/bitcoin/bitcoin/pull/21178 | core#21178]].

Depends on D12434.

Note that in the source material the locktime is not passed down from
send_self_transfer to create_self_transfer which is an obvious oversight.
This is silently fixed in [[https://github.com/bitcoin/bitcoin/pull/22210 | core#22210]] but I also fixed it in this diff.

Test Plan:
  ./test/functional/test_runner.py mempool_reorg

Reviewers: #bitcoin_abc, sdulfari

Reviewed By: #bitcoin_abc, sdulfari

Subscribers: sdulfari

Differential Revision: https://reviews.bitcoinabc.org/D12435