[avalanche] Download the proof from our peers that send avahello
Summary:
This makes the `avahello` message to behave as an equivalent to sending
a proof inventory. If we don't already know about the proof, the proof
id is added to the tracker so it can be requested for download.

Depends on D9625.

Part of T1611.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Subscribers: baby636

Maniphest Tasks: T1611

Differential Revision: https://reviews.bitcoinabc.org/D9627