[avalanche] Add an unbroadcast set of proofs
Summary:
When a proof is manually submitted via sendavalancheproof, an inventory
is sent to all the peers but they are not tracked like other proof invs.
An unbroadcasted set of proofs is added to the peermanager that allow
for reattempt broadcasting these proofs until a peer has requested it.
This is similar to what is done with transactions, and will also make it
possible to store our proofs to a file and broadcast it at startup like
we do for the mempool content.

Ref T1611.

Test Plan:
  ninja check-functional

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Subscribers: PiRK

Maniphest Tasks: T1611

Differential Revision: https://reviews.bitcoinabc.org/D9688