[avalanche] Keep a rolling bloom filter of our recently announced proofs
Summary:
Similar to what is done for the transactions, this will be used to
filter which proof can be requested for download by a peer. The intent
is to improve privacy by preventing peers from being able to guess our
proof pool. The rolling bloom filter parameters are kept identical to
the ones used by the transactions, which itself is based on the
inventory rate limit.

There is no change in behavior at this stage.

Ref T1611.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Maniphest Tasks: T1611

Differential Revision: https://reviews.bitcoinabc.org/D9642