[avalanche] Keep track of the peer creation time
Summary:
Similar to what is done for the transactions, this will be used to filter which proof a peer can request, by only allowing for download the proofs which are in our pool for long enough.
The intent is to improve privacy by preventing peers from being able to guess our proof pool. It is a complementary feature to be used in conjunction with D9642.

There is no change in behavior at this stage.

Ref T1611.

Test Plan:
  ninja all check-avalanche

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Subscribers: deadalnix

Maniphest Tasks: T1611

Differential Revision: https://reviews.bitcoinabc.org/D9643