[avalanche] Move the global definitions into their own header
Summary:
Move the declarations that belong to the global namespace into their own
header. This reduce the scope of inclusion when all that is needed is to
check the -enableavalanche option, and will also prevent from
introducing circular dependencies between the avalanche processor and
the network layer.

Ref T1611.

Test Plan:
  ninja all check

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Maniphest Tasks: T1611

Differential Revision: https://reviews.bitcoinabc.org/D9647