[avalanche] Check the return value of the delegation addLevel() at startup
Summary:
Because the key mismatch is already checked, this can only fail if the Schnorr signature fails.
There is no way to trigger this from a test, but in any case we should always check the return to make the init code more robust.

Depends on D9365.

Ref T1635.

Test Plan:
  ninja check-functional

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Maniphest Tasks: T1635

Differential Revision: https://reviews.bitcoinabc.org/D9668