[avalanche] Make sure the master key matches the proof master
Summary:
As per title. This would cause the `addLevel()` to fail but only if the
session key mismatch the proof master. Adding an explicit test for this
case will always catch the issue. This will be a better solution than
D9668 when we will start to supply delegations as well.

Ref T1635.

Test Plan:
  ./test/functional/test_runner.py abc_rpc_avalancheproof

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Maniphest Tasks: T1635

Differential Revision: https://reviews.bitcoinabc.org/D9669