validation: add chainman ref to CChainState
Summary:
Add an upwards reference to chainstate instances to the owning
ChainstateManager. This is necessary because there are a number
of `this_chainstate == chainman.ActiveChainstate()` checks that
will happen (as a result of assumeutxo) in functions that otherwise
don't have an easily-accessible reference to the chainstate's
ChainManager.

This is a backport of [[https://github.com/bitcoin/bitcoin/pull/21526 | core#21526]] [1/12]
https://github.com/bitcoin/bitcoin/pull/21526/commits/9f6bb539359b98d5b39482ab8a28a68608f0c645

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D12277