[avalanche] Check the delegation generated at startup is valid
Summary:
This makes sure we don't generate an invalid delegation. The proof and
delegation verification are moved to their own function to unbloat the
processor factory.

The appropriated tests will be added when the `-avadelegation` is added as it will allow for triggering the failures.

Ref T1635.

Test Plan:
  ninja check-functional

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Subscribers: deadalnix

Maniphest Tasks: T1635

Differential Revision: https://reviews.bitcoinabc.org/D9445